Fix errors (first start) #1

+++ Source/Game/Conditions/Condition.cs
@@ -319,7 +319,7 @@ namespace Game.Conditions
                     break;
                 case ConditionTypes.Title:
                     if (player != null)
-                        condMeets = player.HasTitle(ConditionValue1);
+                        condMeets = player.HasTitle((int)ConditionValue1);
                     break;
                 case ConditionTypes.UnitState:
                     if (unit != null)
